requirements.txt
# RAG Chatbot Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.3.0                    # OpenAI API client
gradio>=4.0.0                   # Web UI framework
numpy>=1.24.0                   # Numerical operations
faiss-cpu>=1.7.4               # Vector similarity search (CPU version)

# PDF processing
PyMuPDF>=1.23.0                # PDF text and image extraction
pdfplumber>=0.10.0             # Alternative PDF processor
Pillow>=10.0.0                 # Image processing

# OCR engines (choose one or both)
easyocr>=1.7.0                 # Easy OCR engine (recommended)
pytesseract>=0.3.10            # Tesseract OCR wrapper

# Text processing
nltk>=3.8.1                    # Natural language processing
regex>=2023.0.0                # Advanced regex support

# Utilities
pathlib2>=2.3.7               # Path handling (for older Python versions)
tqdm>=4.65.0                   # Progress bars
python-dotenv>=1.0.0          # Environment variable loading

# Optional: Enhanced text processing
# spacy>=3.6.0                 # Advanced NLP (uncomment if needed)
# transformers>=4.30.0         # Hugging Face transformers (for advanced features)

# Development and debugging
loguru>=0.7.0                  # Enhanced logging
click>=8.1.0                   # Command line interface

# Data handling
pandas>=2.0.0                  # Data manipulation
json5>=0.9.0                   # JSON with comments support

# System requirements notes:
# - Python 3.8+ required
# - For GPU acceleration, replace faiss-cpu with faiss-gpu
# - Tesseract binary must be installed separately for pytesseract
# - EasyOCR will download models automatically on first use

